Restart with write/compute potential work now in parallel
The fields were only added on the processor that hosted the grid.
Needed to move this code outside of the local if-statement.  Because I
modeled it after the code adding the MHD fields, I assume that this
did not work in parallel either.  So I also moved them outside of the
MyProcessorNumber == ProcessorNumber if-block.